chore: add additional writable directory to openzaak celery
